bsp: linux-lmp-ti-staging: move RDEPENDS and RRECOMMENDS from kernel-rdepends.inc
Instead of making kernel RDEPENDS on devicetree and different FW images with
lots of machine-override appends that are hard to modify downstream, use proper
MACHINE_ESSENTIAL_EXTRA_RDEPENDS and MACHINE_ESSENTIAL_EXTRA_RRECOMMENDS vars
that are specifically designed for this purpose.

This also makes such dependencies generic and not tied to a particular kernel
recipe including kernel-rdepends.inc with the entire machine-specific list.

https://git.yoctoproject.org/meta-ti/commit/?id=ef80a01ba0294d9f9c045c252ecef7883198b0cb

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>